requirements.txt
fastavro
httpx
aiofiles

# Code quality
# ------------------------------------------------------------------------------
black
flake8
mypy
isort

# Testing
# ------------------------------------------------------------------------------
faker
pytest-asyncio

# Docs
# ------------------------------------------------------------------------------
mkdocs
mkdocs-material

# Coverage
#-------------------------------------------------------------------------------
codecov
pytest-cov
